update sdk
6.2 git version: 369e20401a
1. [AmebaD][Release][auto] integrated patch release
2. update release note
3. [amebad][i2c] add timeout for usi-i2c
4. update release note
5. [amebad][wlan][halrf][bug fix-normal] Fix tracking issue
6. [amebad][bt:ble] Fix coverity high impact items BT part
7. [amebad][wlan] add API to get softap client information
8. [amebad][bt:controller] Update BT normal patch
9. [amebad][wlan] add customization for HP of countries channel plan
10. [AmebaD][at_cmd] Using AT+HTTPCLIENT, need to configure ATVER_2, AT command response should be print on MCU UART
11. [AmebaD][at_cmd] Fix the hardfault issue about AT+HTTPCLIENT
12. [amebad][wlan]disable extended countryinfo after scan done
13. [amebad][bt:mesh]update mesh lib to v2.4.0
14. [amebaD][WebSocket] fix WebSocket server sending masked frame
15. [amebaD][dma] fix bug that allocate dma channel conflict
16. [amebaD][uart] update mbed api
17. [amebad][wlan] add print logs to display country code and channel plan
18. [amebad][wlan] Reduce probe response when receiving probe request continuously
19. [amebaD][mbedtls] Modification to align the struct size in mbedtls 3.6.0
20. [amebad][example] add wifi raw example
21. [amebad][wlan] fix DHCPS receive packages from wlan0 in concurrent mode
22. [amebad][wlan] fix getting wrong ip addr with recv multi-dns enabling
23. [amebaD][wlan] Extend scan timeout for con-current mode
24. [amebad][matter] support matter part 4
25. [amebad][github] remove usi related files
26. [amebad][github] add matter_setup.sh to release list

+++ component/common/network/ssl/mbedtls-3.6.0/include/mbedtls/bignum.h
@@ -222,10 +222,10 @@ typedef struct mbedtls_mpi {
      * Note that this implies that calloc() or `... = {0}` does not create
      * a valid MPI representation. You must call mbedtls_mpi_init().
      */
-    signed short MBEDTLS_PRIVATE(s);
+    int MBEDTLS_PRIVATE(s);
 
     /** Total number of limbs in \c p.  */
-    unsigned short MBEDTLS_PRIVATE(n);
+    size_t MBEDTLS_PRIVATE(n);
     /* Make sure that MBEDTLS_MPI_MAX_LIMBS fits in n.
      * Use the same limit value on all platforms so that we don't have to
      * think about different behavior on the rare platforms where
